+++ mandelimg.h
@@ -1,4 +1,4 @@
 #ifndef MANDELIMG_H
 #define MANDELIMG_H
-void drawMandelbrot(unsigned int,unsigned int,unsigned int,double,double,double);
+void drawMandelbrot(unsigned int,unsigned int,unsigned int,__float128,__float128,__float128);
 #endif
